changed && added: badge.tsx to now seperate the imports of badVariants due to hot refreshing ( refreshing components, i.e only the edited fields in a document rather than the entire page ) only being able to work with exported components, following few commits will follow this exact same format as they're all the same thing, so i will not be writing out a message for each of those icl
