+++ frontend/@/components/ui/badge.tsx
@@ -1,31 +1,36 @@
-import { cva, type VariantProps } from "class-variance-authority"
-import { Slot } from "radix-ui"
-import * as React from "react"
+// import { cva, type VariantProps } from "class-variance-authority"
+import { badgeVariants } from "./badge-variants";
+import type { VariantProps } from "class-variance-authority";
 
-import { cn } from "@/lib/utils"
+import { Slot } from "radix-ui";
+import * as React from "react";
 
-const badgeVariants = cva(
-  "group/badge inline-flex h-5 w-fit shrink-0 items-center justify-center gap-1 overflow-hidden rounded-3xl border border-transparent px-2 py-0.5 text-xs font-medium whitespace-nowrap transition-all focus-visible:border-ring focus-visible:ring-[3px] focus-visible:ring-ring/50 has-data-[icon=inline-end]:pr-1.5 has-data-[icon=inline-start]:pl-1.5 aria-invalid:border-destructive aria-invalid:ring-destructive/20 dark:aria-invalid:ring-destructive/40 [&>svg]:pointer-events-none [&>svg]:size-3!",
-  {
-    variants: {
-      variant: {
-        default: "bg-primary text-primary-text  [a]:hover:bg-primary/80",
-        secondary:
-          "bg-secondary text-secondary-text [a]:hover:bg-secondary/80",
-        destructive:
-          "bg-destructive/10 text-destructive focus-visible:ring-destructive/20 dark:bg-destructive/20 dark:focus-visible:ring-destructive/40 [a]:hover:bg-destructive/20",
-        outline:
-          "border-border text [a]:hover:bg-muted [a]:hover:text-muted-foreground",
-        ghost:
-          "hover:bg-muted hover:text-muted-foreground dark:hover:bg-muted/50",
-        link: "text-primary underline-offset-4 hover:underline",
-      },
-    },
-    defaultVariants: {
-      variant: "default",
-    },
-  }
-)
+import { cn } from "@/lib/utils";
+
+// const badgeVariants = cva(
+//   "group/badge inline-flex h-5 w-fit shrink-0 items-center justify-center gap-1 overflow-hidden rounded-3xl border border-transparent px-2 py-0.5 text-xs font-medium whitespace-nowrap transition-all focus-visible:border-ring focus-visible:ring-[3px] focus-visible:ring-ring/50 has-data-[icon=inline-end]:pr-1.5 has-data-[icon=inline-start]:pl-1.5 aria-invalid:border-destructive aria-invalid:ring-destructive/20 dark:aria-invalid:ring-destructive/40 [&>svg]:pointer-events-none [&>svg]:size-3!",
+//   {
+//     variants: {
+//       variant: {
+//         default: "bg-primary text-primary-text  [a]:hover:bg-primary/80",
+//         secondary:
+//           "bg-secondary text-secondary-text [a]:hover:bg-secondary/80",
+//         destructive:
+//           "bg-destructive/10 text-destructive focus-visible:ring-destructive/20 dark:bg-destructive/20 dark:focus-visible:ring-destructive/40 [a]:hover:bg-destructive/20",
+//         outline:
+//           "border-border text [a]:hover:bg-muted [a]:hover:text-muted-foreground",
+//         ghost:
+//           "hover:bg-muted hover:text-muted-foreground dark:hover:bg-muted/50",
+//         link: "text-primary underline-offset-4 hover:underline",
+//       },
+//     },
+//     defaultVariants: {
+//       variant: "default",
+//     },
+//   }
+// )
+//
+export { Badge };
 
 function Badge({
   className,
@@ -34,16 +39,18 @@ function Badge({
   ...props
 }: React.ComponentProps<"span"> &
   VariantProps<typeof badgeVariants> & { asChild?: boolean }) {
-  const Comp = asChild ? Slot.Root : "span"
+  const Comp = asChild ? Slot.Root : "span";
 
   return (
     <Comp
       data-slot="badge"
       data-variant={variant}
-      className={cn(badgeVariants({ variant }), "font-badge font-semibold text-l", className)}
+      className={cn(
+        badgeVariants({ variant }),
+        "font-badge font-semibold text-l",
+        className,
+      )}
       {...props}
     />
-  )
+  );
 }
-
-export { Badge, badgeVariants }
